// src/modules/predictions/predictions.types.ts
export type ValueBet = {
  market: string;
  bet: string;
  bookmaker: string;
  odd: number;
  fair_odd: number;
  stake: number;
  is_value: boolean;
};

export type FixtureValueBetsResponse = {
  fixture_id: number;
  bets: ValueBet[];
};

export type RawPrediction = {
  id: number;
  fixture_id: number;
  type_id: number;
  predictions: Record<string, any>;
};